requirements.txt
pandas==2.2.3
matplotlib==3.8.2
openpyxl==3.1.2
xlrd==2.0.1
pyyaml==6.0.2